+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add basic support for Numpydoc and reStructuredText docstrings.
    The most common rST elements are supported, but we do not intend
    to support the full complexity of the spec.
+ - Links within the current module now don't require the full qualified path.
  - Improvements to the default theme.
 
 # 2021-01-22: pdoc 2.0.0